pyproject.toml
[tool.poetry]
name = "ualr-chatbot"
version = "0.1.0"
description = ""
authors = ["Ishmam <ishmam.a.solaiman@gmail.com>"]
packages = [
  { include = "app" }
]


[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
faiss-cpu = "^1.10.0"
fastapi = "^0.115.12"
uvicorn = "^0.34.2"
google-genai = "^1.14.0"
langchain-google-genai = "^2.1.4"
pandas = "^2.2.2"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
notebook = "^7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
